Refocus AFL stats integration as one-time historical import
The afltables adapter was overbuilt for its actual use case. Primary use
is a one-time historical backfill (dev tool), not a recurring production
integration. A possible future weekly reconciliation use case is YAGNI
until confirmed.

Removed:
- afltables adapter, cache, doc (weekly production concerns)
- application/ports.go (StatsProvider port — not needed without adapter)
- dev/postgres/init/03_afl_integrations.sql (xref table — will be
  restored when the import tool is built)

Kept and simplified:
- parser.go — PlayerMatchStats struct moved here (owns its output type)
- clubs.go — team code → club name mapping

Docs:
- New ai/architecture/historical-import.md — two-phase reconcile →
  import pattern, Levenshtein fuzzy matching, xref-per-source, FFL
  parallel, committed reconcile.csv artifact
- integrations.md one-time imports section now references the new doc
- Roadmap: phases 16–19 added for AFL historical data (three sources
  covering 2024–1998) and FFL historical data; Deployment renumbered 20
- Sprint updated to reflect the actual shape of the work
